requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
